Move uncached assets out of common
I'm trying to get rid of `/common` so I'm moving all the assets related logic out of `common` and inside `facia-tool`.
