+++ common/app/common/package.scala
@@ -5,7 +5,6 @@ import java.util.concurrent.TimeoutException
 import akka.pattern.CircuitBreakerOpenException
 import com.gu.contentapi.client.GuardianContentApiError
 import conf.SwitchTrait
-import model.{Cached, NoCache}
 import play.api.Logger
 import play.api.libs.json.{JsObject, JsString}
 import play.api.mvc.{RequestHeader, Result}
@@ -16,22 +15,22 @@ object `package` extends implicits.Strings with implicits.Requests with play.api
   def convertApiExceptions[T](implicit log: Logger): PartialFunction[Throwable, Either[T, Result]] = {
     case e: CircuitBreakerOpenException =>
       log.error(s"Got a circuit breaker open error while calling content api")
-      Right(NoCache(ServiceUnavailable))
+      Right(ServiceUnavailable)
     case GuardianContentApiError(404, message) =>
       log.info(s"Got a 404 while calling content api: $message")
-      Right(NoCache(NotFound))
+      Right(NotFound)
     case GuardianContentApiError(410, message) =>
       log.info(s"Got a 410 while calling content api: $message")
-      Right(NoCache(Gone))
+      Right(Gone)
     case timeout: TimeoutException =>
       log.info(s"Got a timeout while calling content api: ${timeout.getMessage}")
-      Right(NoCache(GatewayTimeout(timeout.getMessage)))
+      Right(GatewayTimeout(timeout.getMessage))
     case error =>
       log.info(s"Content api exception: ${error.getMessage}")
       Option(error.getCause).map { cause =>
         log.info(s"Content api exception cause: ", cause)
       }
-      Right(NoCache(InternalServerError))
+      Right(InternalServerError)
   }
 
   /*
@@ -39,15 +38,15 @@ object `package` extends implicits.Strings with implicits.Requests with play.api
           Only the once you actually render is used
    */
 
-  def renderFormat(htmlResponse: () => Html, jsonResponse: () => Html, metaData: model.MetaData)(implicit request: RequestHeader) = Cached(metaData) {
+  def renderFormat(htmlResponse: () => Html, jsonResponse: () => Html, metaData: model.MetaData)(implicit request: RequestHeader) = {
     Ok(htmlResponse())
   }
 
-  def renderFormat(htmlResponse: () => Html, jsonResponse: () => Html, metaData: model.MetaData, switches: Seq[SwitchTrait])(implicit request: RequestHeader) = Cached(metaData) {
+  def renderFormat(htmlResponse: () => Html, jsonResponse: () => Html, metaData: model.MetaData, switches: Seq[SwitchTrait])(implicit request: RequestHeader) = {
     Ok(htmlResponse())
   }
 
-  def renderFormat(htmlResponse: () => Html, jsonResponse: () => Html, cacheTime: Integer)(implicit request: RequestHeader) = Cached(cacheTime) {
+  def renderFormat(htmlResponse: () => Html, jsonResponse: () => Html, cacheTime: Integer)(implicit request: RequestHeader) = {
     Ok(htmlResponse())
   }
