+++ facia-tool/app/controllers/UncachedAssets.scala
@@ -1,3 +1,7 @@
-package controllers.faciatool
+package controllers
 
-object UncachedAssets extends controllers.PublicAssets
+object UncachedAssets extends play.api.mvc.Controller {
+  def at(file: String, relativePath: String = "") = model.NoCache {
+    controllers.Assets.at("/public", relativePath + file)
+  }
+}
