+++ facia-tool/app/views/admin_main.scala.html
@@ -9,8 +9,8 @@
         <title>Fronts Editor</title>
         @templates.meta()
 
-        <script src="@controllers.faciatool.routes.UncachedAssets.at("jspm_packages/system.js")"></script>
-        <script src="@controllers.faciatool.routes.UncachedAssets.at("js/jspm-config.js")"></script>
+        <script src="/assets/jspm_packages/system.js"></script>
+        <script src="/assets/js/jspm-config.js"></script>
 
         <script>System.import('setup');</script>
     </head>
