+++ facia-tool/app/views/templates/header.scala.html
@@ -2,7 +2,7 @@
 
 <header>
     <a href="/" target="_top">
-        <img class="logo" src="@controllers.faciatool.routes.UncachedAssets.at("images/edit-pen.png")" />
+        <img class="logo" src="/assets/images/edit-pen.png" />
         <h1>
             <span data-bind="text: title()">fronts</span>
             @if(Play.isDev && stage.toLowerCase != "dev") { : @stage }
