+++ facia-tool/app/views/templates/meta.scala.html
@@ -1,16 +1,16 @@
 <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
 <meta charset="utf-8"/>
 
-<link rel="shortcut icon" type="image/png" href="@controllers.faciatool.routes.UncachedAssets.at("images/favicon.ico")" />
+<link rel="shortcut icon" type="image/png" href="/assets/images/favicon.ico" />
 
-<link rel="apple-touch-icon" sizes="144x144" href="@controllers.faciatool.routes.UncachedAssets.at("images/144x144.png")" />
-<link rel="apple-touch-icon" sizes="114x114" href="@controllers.faciatool.routes.UncachedAssets.at("images/114x114.png")" />
-<link rel="apple-touch-icon" sizes="72x72" href="@controllers.faciatool.routes.UncachedAssets.at("images/72x72.png")" />
-<link rel="apple-touch-icon-precomposed" href="@controllers.faciatool.routes.UncachedAssets.at("images/57x57.png")" />
+<link rel="apple-touch-icon" sizes="144x144" href="/assets/images/144x144.png" />
+<link rel="apple-touch-icon" sizes="114x114" href="/assets/images/114x114.png" />
+<link rel="apple-touch-icon" sizes="72x72" href="/assets/images/72x72.png" />
+<link rel="apple-touch-icon-precomposed" href="/assets/images/57x57.png" />
 
 <meta name="apple-mobile-web-app-title" content="Guardian" />
 <meta name="application-name" content="The Guardian" />
 <meta name="msapplication-TileColor" content="#005689" />
-<meta name="msapplication-TileImage" content="@controllers.faciatool.routes.UncachedAssets.at("images/windows_tile_144_b.png")" />
+<meta name="msapplication-TileImage" content="/assets/images/windows_tile_144_b.png" />
 
-<link href="@controllers.faciatool.routes.UncachedAssets.at("css/style.css")" rel="stylesheet">
+<link href="/assets/css/style.css" rel="stylesheet">
