+++ facia-tool/app/views/unauthorized.scala.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="utf-8">
         <title>Unauthorized - Fronts Editor</title>
-        <link href="@controllers.faciatool.routes.UncachedAssets.at("css/style.css")" rel="stylesheet">
+        <link href="/assets/css/style.css" rel="stylesheet">
     </head>
     <body class="unauthorizedMessage">
         <div class="frownFace">
